+++ R/imports.R
@@ -1,3 +1,3 @@
 #' @importFrom graphics axis legend points polygon segments
-#' @importFrom stats terms
+#' @importFrom stats terms as.formula aggregate
 NULL