+++ NEWS.md
@@ -4,3 +4,4 @@ Lexiguel 0.1.0
 ### New features
 
 * New function `tab4net()` to process data frames for `sankeyNetwork()`.
+* New function `lim_bars()` to produce limits bar plots.